requirements.txt
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
Django==5.1.4
django-eventstream==5.3.1
django-grip==3.5.1
gripcontrol==4.2.0
h11==0.14.0
httptools==0.6.4
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
MarkupSafe==3.0.2
mysqlclient==2.2.6
pubcontrol==3.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.3.0
requests==2.32.3
scapy==2.6.1
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
sseclient-py @ file:///D:/stephenzhu/Desktop/%E8%AE%A1%E7%AE%97%E6%9C%BA%E7%BD%91%E7%BB%9C%E5%A4%A7%E4%BD%9C%E4%B8%9A/Network-Sniffer/client/requirements/sseclient_py-1.8.0-py2.py3-none-any.whl#sha256=4ecca6dc0b9f963f8384e9d7fd529bf93dd7d708144c4fb5da0e0a1a926fee83
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
zope.interface==7.2
